+++ src/index.js
@@ -1,59 +1,48 @@
-import { Router } from "flight-path";
+import { includeBytes } from "fastly:experimental";
+import { Router } from "@fastly/expressly";
 import {
   getGraphQLParameters,
   processRequest,
   renderGraphiQL,
   shouldRenderGraphiQL,
-  sendResult,
 } from "graphql-helix";
 import { schema } from "./schema";
 
 const router = new Router();
 
-import ManifestContent from "./assets/manifest.yaml";
-router.route("*", "/.well-known/fastly/demo-manifest", (req, res) => {
-  res.setHeader("Content-Type", "text/plain");
-  res.send(ManifestContent)
-})
+import ManifestContent from "./assets/manifest.md";
+router.get("/.well-known/fastly/demo-manifest", (_, res) => {
+  res.text(ManifestContent);
+});
 
-const png = fastly.includeBytes("src/assets/graphql-logo.png");
-router.get("/logo.png", (req, res) => {
-  res.setHeader("Content-Type", "image/png");
-  res.send(png)
-})
+const png = includeBytes("src/assets/graphql-logo.png");
+router.get("/logo.png", (_, res) => {
+  res.headers.set("content-type", "image/png");
+  res.send(png);
+});
 
 import logoSvg from "./assets/graphql-logo.svg";
-router.get("/logo.svg", (req, res) => {
-  res.setHeader("Content-Type", "image/svg+xml");
-  res.send(logoSvg)
-})
+router.get("/logo.svg", (_, res) => {
+  res.headers.set("content-type", "image/svg+xml");
+  res.send(logoSvg);
+});
 
-router.route("*", "*", async (req, res) => {
-  let body = {};
-
-  try {
-    body = await req.json();
-  } catch (e) {}
-
-  // Create a generic Request object that can be consumed by Graphql Helix's API
-  const request = {
-    body: body,
-    headers: req.headers,
-    method: req.method,
-    query: req.query,
-  };
-
-  /**
-   * If the request does not include an accept header (or has a non browser like value), for example basic curls, add a default
-   */
-  if (!request.headers.accept || request.headers.accept === "*/*") {
-    request.headers.accept = "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,*/*;q=0.8";
+router.all("*", async (req, res) => {
+  // If the request does not include an accept header (or has a non-browser-like value), 
+  // add a default.
+  if (
+    !req.headers.get("accept") ||
+    req.headers.get("accept") === "*/*"
+  ) {
+    req.headers.set(
+      "accept",
+      "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,*/*;q=0.8"
+    );
   }
 
-  // Determine whether we should render GraphiQL instead of returning an API response
-  if (shouldRenderGraphiQL(request)) {
-    res.setHeader("Content-Type", "text/html");
-    res.send(
+  // Determine whether we should render GraphiQL instead of returning an API response.
+  if (shouldRenderGraphiQL(req)) {
+    res.html(
       renderGraphiQL({
         defaultQuery: `# Enter your GraphQL query here
 query {
@@ -66,10 +55,20 @@ query {
       })
     );
   } else {
-    // Extract the Graphql parameters from the request
+    let body = await req.json().catch(() => ({}));
+
+    // Create a generic Request object that can be consumed by Graphql Helix's API.
+    const request = {
+      body: body,
+      headers: req.headers,
+      method: req.method,
+      query: req.query,
+    };
+  
+    // Extract the Graphql parameters from the request.
     const { operationName, query, variables } = getGraphQLParameters(request);
 
-    // Validate and execute the query
+    // Validate and execute the query.
     const result = await processRequest({
       operationName,
       query,
@@ -78,7 +77,11 @@ query {
       schema,
     });
 
-    sendResult(result, res);
+    // Respond with the result.
+    for (const { name, value } of result.headers) {
+      res.headers.set(name, value);
+    }
+    res.json(result.payload);
   }
 });