+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 ## 2021-07-28
 - Initial release
+
+## 2023-04-03
+– Updated `@fastly/js-compute` and performed associated changes
+- Migrated from `flight-path` to `@fastly/expressly`